// projestoque/src/Components/textInput/Types.ts
export interface IInput {
  fontSize?: number;
  width?: number | string;
  placeholder?: string;
  onChangeText?: (text: string) => void;
  secureTextEntry?: boolean;
  keyboardType?: any;
  marginTop?: number;
  marginBottom?: number;
  value?: number | string | any;
  editable?: boolean;
  BgColor?: string;
  rounded?: boolean;
  heigth?: number | string;
  padL?: number;
  color?: string;
  radius?: number;
}